+++ RCNutilidades/public/javascripts/index_behaviours.js
@@ -1,11 +1,12 @@
-let produtos = require("./produtos");
+const { inserirProduto } = require("./produtos");
+
 let btnadd = document.getElementById("btn-add");
 let lista_estoque = document.getElementById("lista_estoque");
-let codprods = new Map();
 
-btnadd.addEventListener("click", function(){
-    console.log("Testing Script...");
-});
+
+btnadd.addEventListener("click", inserirProduto());
+
+
